modifying delete func checking day

+++ site3/polls/views.py
@@ -127,7 +127,8 @@ def delete_view(request, usr_pk):
             usr = USR.objects.get(pk=usr_pk)
             try:
                 if usr.polluser.user == request.user:
-                    usr.delete()
+                    if usr.section.eventday.day > 99:
+                        usr.delete()
                 return HttpResponseRedirect(reverse('polls:home'))
             except:
                 return HttpResponseRedirect(reverse('polls:home'))
